. Restructured packages, moved Repositories to package, added client and business packages. . Major bug fixes in Vehicle and CustomerVehicleMapping Lifecycles
