+++ src/main/java/ams/client/model/Address.java
@@ -1,0 +1,40 @@
+package ams.client.model;
+
+import ams.client.model.constants.AddressType;
+import ams.client.model.constants.CountryCode;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.ManyToOne;
+import java.util.UUID;
+
+@EqualsAndHashCode(callSuper = true)
+@Entity
+@Data
+public class
+Address extends Persistence {
+    @Id
+    @GeneratedValue
+    private UUID addressId;
+
+    private String address1;
+
+    private String address2;
+
+    private String city;
+
+    private String state;
+
+    private String zipCode;
+
+    private CountryCode countryCode;
+
+    private AddressType addressType;
+
+    @ManyToOne
+    private Customer customer;
+
+}
