+++ src/main/java/ams/client/model/constants/AddressType.java
@@ -1,0 +1,6 @@
+package ams.client.model.constants;
+
+public enum AddressType {
+
+    BILLING, HOME, OFFICE, OTHERS
+}
