+++ src/main/java/ams/client/model/constants/CountryCode.java
@@ -1,0 +1,6 @@
+package ams.client.model.constants;
+
+
+public enum CountryCode {
+    US, CA, UK, IN, FR
+}
