+++ src/main/java/ams/client/repositories/CustomerRepository.java
@@ -1,0 +1,13 @@
+package ams.client.repositories;
+
+import ams.client.model.Customer;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
+
+@Repository
+public interface CustomerRepository extends CrudRepository<Customer, UUID> {
+
+    Customer findCustomerByEmail(String email);
+}
