+++ src/main/java/ams/client/services/CustomerService.java
@@ -1,0 +1,23 @@
+package ams.client.services;
+
+import ams.client.model.Customer;
+import ams.client.repositories.CustomerRepository;
+import lombok.Data;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+@Service
+@Data
+public class CustomerService {
+
+    @Autowired
+    private CustomerRepository customerRepository;
+
+    public Customer searchByEmail(String email) {
+        return customerRepository.findCustomerByEmail(email);
+    }
+
+    public void addCustomer(Customer customer) {
+        customerRepository.save(customer);
+    }
+}
