--- src/main/java/com/abc/ams/AmsApplication.java
@@ -1,19 +1,0 @@
-package com.abc.ams;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
-
-@SpringBootApplication
-@EnableTransactionManagement
-@EnableJpaRepositories
-@EnableJpaAuditing
-public class AmsApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AmsApplication.class, args);
-    }
-
-}
