--- src/main/java/com/abc/ams/CustomerRepository.java
@@ -1,13 +1,0 @@
-package com.abc.ams;
-
-import com.abc.ams.model.Customer;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface CustomerRepository extends CrudRepository<Customer, UUID> {
-
-    Customer findCustomerByEmail(String email);
-}
