--- src/main/java/com/abc/ams/model/Persistence.java
@@ -1,21 +1,0 @@
-package com.abc.ams.model;
-
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-
-import javax.persistence.Column;
-import javax.persistence.MappedSuperclass;
-import java.io.Serializable;
-
-@MappedSuperclass
-public class Persistence implements Serializable {
-
-    @Column(name = "created_date", nullable = false, updatable = false)
-    @CreatedDate
-    private long createdDate;
-
-    @Column(name = "modified_date")
-    @LastModifiedDate
-    private long modifiedDate;
-
-}
