--- src/main/java/com/abc/ams/model/constants/AddressType.java
@@ -1,6 +1,0 @@
-package com.abc.ams.model.constants;
-
-public enum AddressType {
-
-    BILLING, HOME, OFFICE, OTHERS
-}
