--- src/main/java/com/abc/ams/model/constants/CountryCode.java
@@ -1,6 +1,0 @@
-package com.abc.ams.model.constants;
-
-
-public enum CountryCode {
-    US, CA, UK, IN, FR
-}
