+++ src/test/java/ams/AmsApplicationTests.java
@@ -1,0 +1,13 @@
+package ams;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AmsApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
