--- src/test/java/com/abc/ams/AmsApplicationTests.java
@@ -1,13 +1,0 @@
-package com.abc.ams;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AmsApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
